+++ R/s2.R
@@ -16,7 +16,7 @@ sf_use_s2 = function(use_s2) {
 		if (use_s2 && !requireNamespace("s2", quietly = TRUE))
 			stop("package s2 not available: install it first?")
 		if (ret_val != use_s2)
-			cat(paste0("Spherical geometry (s2) switched ", ifelse(use_s2, "on", "off"), "\n"))
+			message(paste0("Spherical geometry (s2) switched ", ifelse(use_s2, "on", "off")))
 		assign(".sf.use_s2", use_s2, envir = .sf_cache)
 		invisible(ret_val)
 	} else